Conflicts:
	src/main/scala/com/leagueprojecto/api/JsonProtocols.scala
	src/main/scala/com/leagueprojecto/api/Routes.scala
	src/main/scala/com/leagueprojecto/api/services/riot/MatchService.scala
	src/main/scala/com/leagueprojecto/api/services/riot/RecentMatchesService.scala
	src/main/scala/com/leagueprojecto/api/services/riot/RiotService.scala
	src/main/scala/com/leagueprojecto/api/services/riot/SummonerService.scala

+++ src/main/scala/com/leagueprojecto/api/domain/PlayerStats.scala
@@ -1,8 +1,8 @@
 package com.leagueprojecto.api.domain
 
 case class PlayerStats(
-                        minionKills: Int,
-                        kills: Int,
-                        deaths: Int,
-                        assists: Int
+                        minionKills: Long,
+                        kills: Long,
+                        deaths: Long,
+                        assists: Long
                         )
